fix: use SplittedGrayAddress
